Today `GroupRegistry` employs thread isolation by default, i.e. every thread sees its own process group registry. This is intended to work for one-device-per-process (for python use cases) and one-device-per-thread case (for custom native runtimes).

However, there's a problem - there are python use cases that initializes/registers process groups in one thread, and runs collectives in another thread. This use case should be supported. However, since `GroupRegistry` employs thread isolation by default, collectives in different threads can't find the registered process groups.

This PR fixes the issue by:
- Make `GroupRegistry` work in non-thread isolation mode by default. This would match the behavior w/o the native process group registry.
- Introduces `set_thread_isolation_mode` so one-device-per-thread runtimes can enable thread isolation mode explicitly.

cc mrshenli pritamdamania87 zhaojuanmao satgera rohan-varma gqchen aazzolini osalpekar jiayisuse H-Huang kwen2501 awgu penguinwu fegin XilunWu wanchaol fduwjj wz337 tianyu-l wconstab yf225 chauhang

Differential Revision: [D54658515](https://our.internmc.facebook.com/intern/diff/D54658515)

[ghstack-poisoned]

+++ torch/csrc/jit/runtime/static/impl.h
@@ -938,9 +938,9 @@ class TORCH_API ProcessedNode {
     return values_[outputs_offset_ + i];
   }
 
-  uint32_t num_outputs() const {
+  size_t num_outputs() const {
     DCHECK(fn_ != nullptr);
-    return static_cast<uint32_t>(fn_->num_outputs());
+    return fn_->num_outputs();
   }
 
   C10_NODISCARD c10::ArrayRef<const IValue> outputs() const {
